pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otonom-trader"
version = "0.1.0"
description = "Multi-analyst autonomous trading system with broker abstraction and risk guardrails"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Asim", email = "asim@example.com" }
]
keywords = ["trading", "algorithmic-trading", "crypto", "binance", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "typer>=0.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.14.0",
    "streamlit>=1.22.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
]

[project.scripts]
otonom-trader = "otonom_trader.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/anka_trader"
Documentation = "https://github.com/yourusername/anka_trader/blob/main/README.md"
Repository = "https://github.com/yourusername/anka_trader"
Issues = "https://github.com/yourusername/anka_trader/issues"

[tool.setuptools]
packages = ["otonom_trader"]

[tool.setuptools.package-data]
otonom_trader = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=otonom_trader --cov-report=html --cov-report=term-missing"
